:building_construction: Introduced controller logic for AST
