+++ src/test/java/za/ac/sun/grapl/intraprocedural/ArithmeticTest.java
@@ -11,6 +11,7 @@ import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.TestInfo;
 import za.ac.sun.grapl.Cannon;
+import za.ac.sun.grapl.controllers.ASTController;
 import za.ac.sun.grapl.domain.enums.EdgeLabels;
 import za.ac.sun.grapl.domain.models.vertices.BlockVertex;
 import za.ac.sun.grapl.domain.models.vertices.LiteralVertex;
@@ -53,6 +54,7 @@ public class ArithmeticTest {
 
     @BeforeEach
     public void setUp(TestInfo testInfo) throws IOException {
+        ASTController.getInstance().resetOrder();
         TinkerGraphHook hook = new TinkerGraphHook.TinkerGraphHookBuilder(TEST_DIR).createNewGraph(true).build();
         Cannon fileCannon = new Cannon(hook);
         // Select test resource based on integer in method name
