+++ src/test/java/za/ac/sun/grapl/intraprocedural/BasicIntraproceduralTest.java
@@ -12,6 +12,7 @@ import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.TestInfo;
 import za.ac.sun.grapl.Cannon;
+import za.ac.sun.grapl.controllers.ASTController;
 import za.ac.sun.grapl.domain.enums.EdgeLabels;
 import za.ac.sun.grapl.domain.enums.VertexLabels;
 import za.ac.sun.grapl.domain.models.vertices.BlockVertex;
@@ -53,6 +54,7 @@ public class BasicIntraproceduralTest {
 
     @BeforeEach
     public void setUp(TestInfo testInfo) throws IOException {
+        ASTController.getInstance().resetOrder();
         TinkerGraphHook hook = new TinkerGraphHook.TinkerGraphHookBuilder(TEST_DIR).createNewGraph(true).build();
         Cannon fileCannon = new Cannon(hook);
         // Select test resource based on integer in method name
